Move tests for exposed functions
